ml-service/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ML and Image Processing
tensorflow>=2.20.0
numpy>=1.24.3
Pillow>=10.2.0

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
